refactor: encapsulate spy service in chat service / command tree
